+++ modules/features/search-screen/src/main/java/com/test/search_screen/SearchRouter.kt
@@ -2,5 +2,5 @@ package com.test.search_screen
 
 interface SearchRouter {
 
-    fun openObjectDetailsFragment(objectId: Long)
+    fun openObjectDetailsFragment(objectId: Int)
 }
