+++ modules/features/search-screen/src/main/java/com/test/search_screen/adapter/uiItems/SearchResultUiItem.kt
@@ -1,5 +1,5 @@
 package com.test.search_screen.adapter.uiItems
 
-data class SearchResultUiItem(val objectId: Long) : UiItem {
+data class SearchResultUiItem(val objectId: Int) : UiItem {
     override fun getItemId(): Long = System.currentTimeMillis()
 }
